Remove digipog awarding from guests & fix profile permissions
