+++ routes/api/user/delete.js
@@ -1,6 +1,6 @@
 const { logger } = require("../../../modules/logger");
 const { httpPermCheck } = require("../../middleware/permissionCheck");
-const { deleteUser } = require("../../../modules/user/user");
+const { deleteUser } = require("../../../modules/user/userSession");
 
 module.exports = {
     run(router) {
