+++ static/css/style.css
@@ -358,7 +358,6 @@ dialog {
 #muteButton,
 #unmuteButton {
     display: none;
-	margin-right: 0;
 }
 
 #logoutIcon {
